First Fixes on MO 2

+++ src/uniandes/tsdl/mutapk/processors/MutationsProcessor.java
@@ -141,13 +141,13 @@ public class MutationsProcessor {
 						// Perform mutation
 						operator.performMutation(mutationLocation, writer, currentMutationIndex);
 						Long mutationEnd = System.currentTimeMillis();
-						APKToolWrapper.buildAPK(mutantRootFolder, extraPath, apkName, currentMutationIndex);
+						boolean result = APKToolWrapper.buildAPK(mutantRootFolder, extraPath, apkName, currentMutationIndex);
 						File mutatedFile = new File(newMutationPath);
 						String fileName = (new File(newMutationPath)).getName();
 						File mutantRootFolderDir = new File(mutantRootFolder+fileName);
 						FileUtils.copyFile(mutatedFile, mutantRootFolderDir);
 						File srcFolder = new File(mutantFolder);
-						FileUtils.deleteDirectory(srcFolder);
+						if(result) {FileUtils.deleteDirectory(srcFolder);}
 						Long buildEnd = System.currentTimeMillis();
 						Long mutationTime = mutationEnd-mutationIni;
 						Long buildingTime = buildEnd - mutationEnd;
